refactor: generalise log collection in cluster dump script
Make the cluster dump script more generic.
There were too manay specifics in terms of pods and containers,
comprising a mayastor installation.
Change it so that it generates logs for pods and containers,
that are discovered by querying the k8s.
And the immediate benefit of this is that logs for nats pods,
are now generated.

Use this new found capability to generate logs for
pods in the default namespace as well, this is useful
because at the moment test pods exist in the default
namespace.