fix: add support to cleanup in BeforeEach
Currently the test framework is geared to supoport
post-mortem analysis on test failure.
This means the when a test fails, no other tests are run.

However when running for test coverage, this behaviour
prevents meaningful runs of all tests.

In particular when a test has multiple It clauses,
execution of testing for all It clause following a failed
execution of a test for an It clause will fail.

This change makes this behaviour selectable.
If onfail is set to reinstall or reset, then
in the BeforeEach clause, if the resource check fails,
an attempt is made to cleanup resources so that
test can be run meaningfully.