fix: for nightly-stable pipeline, build mayastor docker images and install image
Change nightly-stable pipeline to checkout mayastor and build
docker images and the install image from the checkout.
This guarantees that the tagged mayastor docker images match
the tagged install image
The install image is a construct of and for test.
It contains all the files required by the E2E framework to install
mayastor.