chore: refine handling of cluster checks before each test
When an It clause of a test fails,
typically resources are retained for postmortem analysis.
However this has an adverse effect on subsequent It clauses
as they fail because the cluster is not clean,
typically with reports that pool usage is not 0.
This is an incorrect reason the failure of the It
clause, in particular because in Xray an It clause
maps to a test and test execution.
Attempt to clear this misconception by changing the error
returned to be "unclean test cluster!!!"

Whilst here change the policy for nightly, nightly_full so that the
on before each checks an attempt is made to clean up the
cluster before proceeding with the next It clause,
since with those profiles, clusters are not retained,
and test coverage is more important.