chore: support concurrent runs of tests MQ-1723
Introduce the concept of sessions for test runs
For a session, installation files are created/read/written
to a sub-directory with the name of the session
Similarly the report and log files for a session are generated
in sub-directories under the respective directories
for example for session named ABC,
    log files will be in .../artifacts/logs/ABC
    report files will be in .../artifacts/reports/ABC
    all other files will be under .../artifacts/sessions/ABC

It may seem odd that log files and report files are not
generated under .../artifacts/sessions.
It has been this way to make collection of log and report files
under jenkins simpler.

The default session name is "_noname"

add --session option to
    scripts/e2e-test.sh
add --installroot option to
    scripts/extract-install-image.sh

add sessionDir as configuration item
save resolved configuration file in the session directory
generate install yaml files in the session directory