test(e2e): mq-1871 install and uninstall reports
We can have multiple runs of the install and uninstall tests
per Jenkins run. Reporting multiple install results is
undesirable as each test is (or should be) in an identical
environment. Boil down the results to one report which is a
failure if one exists or a success otherwise.
In the case of the uninstall tests, each may be affected by the
previous test run. Modify each Junit file to specify a class and
description which is associated with the test just run.
The second change will result in N new uninstall tests, where N
is the number of sets of tests run.
Also extended the maximum duration of the self-ci test to 2 hours.