chore: refactor use of the MSV CRD to not use the CRD explicitly MQ-2203
Restructure so that we are not use the MSV CRD structure,
instead use intermediate types.
The will allows us to run the tests using control plane 2 which
does not have MSV CRDs.
Instead the intermediate MSV structs will be filled using the
mayastor kubectl plugin.
Besides the structs, support is required to change the number of
replicas in an MSV, add support function and invoke it instead
of updating the CR in k8s