chore: handle change in control plane installation
The test framework identified the presence of control plane
by checking for the presences of the core-agents the deployment
This has changed, core-agents is now a stateful set
Add the check for presence of core-agents as a stateful as a factor
to auto configure correctly.
Adding a new check and retaining the previous check keeps compatibilty
with previous builds.