chore: ensure a single copy of install bundle in artifacts dir MQ-2404
Extract the install bundle to artifacts/install-bundle/<tag>,
and modify the path configurations to use the new location
for install and uninstall purposes.

When extracting remove previous instance of
    artifacts/install-bundle/<tag>
since no assumption can/should be made regarding the
contents of the same.
This should be okay as long as multiple tests are not run
from the same repo concurrently.