chore: stop using node urls with kubectl-mayastor MQ-2537
With VPN access kubectl-mayastor just works if KUBECONFIG is
set up correctly.

This simplifies the control plane code interfacing with the restful
control-plane.

Rip out most of the offending bits.
But retain the use of IP addresses, to communicate with the
restful control plane for replica information.
Uses master node IP addresses now instead of worker nodes,
this will change when MQ-2319 is merged.