chore: refactor to aid test coverage and development
Add a configuration item "beforeEachCheckAndRestart" to enable
 restarts in the BeforeEachCheck function.
Currently if an 'It' clause of test fails and leaves
resources, the cluster is considered unusable because
of indeterminate state.
Setting beforeEachCheckRestart changes the behaviour such
that mayastor is restarted if
    - If prior ResourceCheck has failed
    - ResourceCheck fails
The 2nd behaviour may help for selfci runs.

Add a configuration item "failQuick" to ease debugging and
development of tests.
When trying to determine reasons for failure, developers
prefer fast fail.
fail quick doesn't quite manage that, it only affects the i
behaviour of the generic BeforeEachCheck and AfterEachCheck functions.
ResourceCheck errors are stored persistently and if fail quick
is enabled both functions return immediately.

Finally change so that on BeforeEachChecks we do not wait
for pool state to reach the desired 0 value
    - the timeouts are large
    - timeouts will have been exercised in the prior
        AfterEachCheck

Refactor to simplify functions which accumulate errors