fix: tweak nexus_location test
After checking whether a pod is provisioned,
sleep for 1 minute and then delete the pod,
this stops the test failing on k8s 1.22 and k8s 1.23.

Ostensibly the test fails because pvc deletion does not
complete successfully, however the root cause is that
the pod using the pvc is not deleted.

Adding the delay lets k8s complete the deletion of the pod,
and naturally pvc deletion will then complete.