fix: use testlists in a consistent manner
use testlists.py and configurations/testlists.yaml in
e2e-test.sh
Create a list of tests that are known to pass
profile name is passing

Before sorting on duration, sort the list of tests alphbetically.
This ensures that the position of tests with the same duration
is the always the same relative to each other.