refactor: handle images for mayastor and bolt
Create install-images for bolt in the registry under datacore/
Fix the fallout from that change.

Make re-tag-images more portable across images.
As a bonus this should just work correctly with older images
which have different image names

predefine variable product to be empty
it is subsequently defaulted to mayastor
this is allows selfci to work a before.
Future changes will fix this,
as we should use bolt images for selfci

re-tag install images as well