refactor: generic-system-test for bolt selfci image build
selfci image build requirements are
 - build
 - test
 - publish as selfci if the previous steps passed

re-jig so that push to dockerhub and publish occur
after the test phases.
generic-system-test is a "library" pipeline, so it does
not error until the final stage.
This makes it possible to generate and acquire artifacts when
tests fail.
A side effect of this is that expclicit tests for failure
are required  before publish and push to dockerhub stages